requirements.txt
torch==1.8.0
torchvision==0.9.0
Pillow==8.1.2
opencv-python==4.5.1.48